Making it support python3
Python3 means that I've had to make it so that when given a unbound
method, it only finds the lineage of the parent that defined it rather
than the parent you are accessing it from.

This is because python3 has no concept of an unbound method and so is no
longer bound to the class you accessed it from.